@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v3.21.12
+// 	protoc-gen-go v1.35.2
+// 	protoc        v3.20.3
 // source: client/rootpb/root.proto
 
 package rootpb
@@ -33,11 +33,9 @@ type Operator struct {
 
 func (x *Operator) Reset() {
 	*x = Operator{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_client_rootpb_root_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_client_rootpb_root_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Operator) String() string {
@@ -48,7 +46,7 @@ func (*Operator) ProtoMessage() {}
 
 func (x *Operator) ProtoReflect() protoreflect.Message {
 	mi := &file_client_rootpb_root_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -103,11 +101,9 @@ type Response struct {
 
 func (x *Response) Reset() {
 	*x = Response{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_client_rootpb_root_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_client_rootpb_root_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Response) String() string {
@@ -118,7 +114,7 @@ func (*Response) ProtoMessage() {}
 
 func (x *Response) ProtoReflect() protoreflect.Message {
 	mi := &file_client_rootpb_root_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -196,7 +192,7 @@ func file_client_rootpb_root_proto_rawDescGZIP() []byte {
 }
 
 var file_client_rootpb_root_proto_msgTypes = make([]protoimpl.MessageInfo, 3)
-var file_client_rootpb_root_proto_goTypes = []interface{}{
+var file_client_rootpb_root_proto_goTypes = []any{
 	(*Operator)(nil), // 0: rootpb.Operator
 	(*Response)(nil), // 1: rootpb.Response
 	nil,              // 2: rootpb.Operator.ParamsEntry
@@ -214,32 +210,6 @@ func init() { file_client_rootpb_root_proto_init() }
 func file_client_rootpb_root_proto_init() {
 	if File_client_rootpb_root_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_client_rootpb_root_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Operator); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_client_rootpb_root_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Response); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
