@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v3.21.12
+// 	protoc-gen-go v1.35.2
+// 	protoc        v3.20.3
 // source: implant/implantpb/implant.proto
 
 package implantpb
@@ -28,11 +28,9 @@ type Empty struct {
 
 func (x *Empty) Reset() {
 	*x = Empty{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_implant_implantpb_implant_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_implant_implantpb_implant_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Empty) String() string {
@@ -43,7 +41,7 @@ func (*Empty) ProtoMessage() {}
 
 func (x *Empty) ProtoReflect() protoreflect.Message {
 	mi := &file_implant_implantpb_implant_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -70,11 +68,9 @@ type Status struct {
 
 func (x *Status) Reset() {
 	*x = Status{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_implant_implantpb_implant_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_implant_implantpb_implant_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Status) String() string {
@@ -85,7 +81,7 @@ func (*Status) ProtoMessage() {}
 
 func (x *Status) ProtoReflect() protoreflect.Message {
 	mi := &file_implant_implantpb_implant_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -134,6 +130,7 @@ type Spite struct {
 	Error   uint32  `protobuf:"varint,5,opt,name=error,proto3" json:"error,omitempty"`
 	Status  *Status `protobuf:"bytes,6,opt,name=status,proto3" json:"status,omitempty"`
 	// Types that are assignable to Body:
+	//
 	//	*Spite_Empty
 	//	*Spite_Block
 	//	*Spite_Ack
@@ -185,11 +182,9 @@ type Spite struct {
 
 func (x *Spite) Reset() {
 	*x = Spite{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_implant_implantpb_implant_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_implant_implantpb_implant_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Spite) String() string {
@@ -200,7 +195,7 @@ func (*Spite) ProtoMessage() {}
 
 func (x *Spite) ProtoReflect() protoreflect.Message {
 	mi := &file_implant_implantpb_implant_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -876,11 +871,9 @@ type Spites struct {
 
 func (x *Spites) Reset() {
 	*x = Spites{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_implant_implantpb_implant_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_implant_implantpb_implant_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Spites) String() string {
@@ -891,7 +884,7 @@ func (*Spites) ProtoMessage() {}
 
 func (x *Spites) ProtoReflect() protoreflect.Message {
 	mi := &file_implant_implantpb_implant_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1132,7 +1125,7 @@ func file_implant_implantpb_implant_proto_rawDescGZIP() []byte {
 }
 
 var file_implant_implantpb_implant_proto_msgTypes = make([]protoimpl.MessageInfo, 4)
-var file_implant_implantpb_implant_proto_goTypes = []interface{}{
+var file_implant_implantpb_implant_proto_goTypes = []any{
 	(*Empty)(nil),                 // 0: implantpb.Empty
 	(*Status)(nil),                // 1: implantpb.Status
 	(*Spite)(nil),                 // 2: implantpb.Spite
@@ -1244,57 +1237,7 @@ func file_implant_implantpb_implant_proto_init() {
 		return
 	}
 	file_implant_implantpb_module_proto_init()
-	if !protoimpl.UnsafeEnabled {
-		file_implant_implantpb_implant_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Empty); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_implant_implantpb_implant_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Status); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_implant_implantpb_implant_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Spite); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_implant_implantpb_implant_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Spites); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
-	file_implant_implantpb_implant_proto_msgTypes[2].OneofWrappers = []interface{}{
+	file_implant_implantpb_implant_proto_msgTypes[2].OneofWrappers = []any{
 		(*Spite_Empty)(nil),
 		(*Spite_Block)(nil),
 		(*Spite_Ack)(nil),
