@@ -8,35 +8,34 @@ const (
 )
 
 const (
-	CtrlTaskCallback     = "task_callback"
-	CtrlTaskFinish       = "task_finish"
-	CtrlTaskCancel       = "task_cancel"
-	CtrlTaskError        = "task_error"
-	CtrlClientJoin       = "client_join"
-	CtrlClientLeft       = "client_left"
-	CtrlWebUpload        = "web_upload"
-	CtrlListenerStart    = "listener_start"
-	CtrlListenerStop     = "listener_stop"
-	CtrlPipelineStart    = "pipeline_start"
-	CtrlPipelineStop     = "pipeline_stop"
-	CtrlWebsiteStart     = "website_start"
-	CtrlWebsiteStop      = "website_stop"
-	CtrlWebsiteRegister  = "website_register"
-	CtrlWebsiteAdd       = "website_add"
-	CtrlJobStart         = "job_start"
-	CtrlJobStop          = "job_stop"
-	CtrlSessionRegister  = "session_register"
-	CtrlSessionDead      = "session_dead"
-	CtrlSessionInit      = "session_init"
-	CtrlSessionReborn    = "session_reborn"
-	CtrlSessionLog       = "session_log"
-	CtrlSessionTask      = "session_task"
-	CtrlSessionError     = "session_error"
-	CtrlSessionLeave     = "session_leave"
-	CtrlSessionCheckin   = "session_checkin"
-	CtrlWebContentAdd    = "web_content_add"
-	CtrlWebContentUpdate = "web_content_update"
-	CtrlWebContentRemove = "web_content_remove"
+	CtrlTaskCallback    = "task_callback"
+	CtrlTaskFinish      = "task_finish"
+	CtrlTaskCancel      = "task_cancel"
+	CtrlTaskError       = "task_error"
+	CtrlClientJoin      = "client_join"
+	CtrlClientLeft      = "client_left"
+	CtrlWebUpload       = "web_upload"
+	CtrlListenerStart   = "listener_start"
+	CtrlListenerStop    = "listener_stop"
+	CtrlPipelineStart   = "pipeline_start"
+	CtrlPipelineStop    = "pipeline_stop"
+	CtrlWebsiteStart    = "website_start"
+	CtrlWebsiteStop     = "website_stop"
+	CtrlWebsiteRegister = "website_register"
+	CtrlRemStart        = "rem_start"
+	CtrlRemStop         = "rem_stop"
+	CtrlRemRegister     = "rem_register"
+	CtrlJobStart        = "job_start"
+	CtrlJobStop         = "job_stop"
+	CtrlSessionRegister = "session_register"
+	CtrlSessionDead     = "session_dead"
+	CtrlSessionInit     = "session_init"
+	CtrlSessionReborn   = "session_reborn"
+	CtrlSessionLog      = "session_log"
+	CtrlSessionTask     = "session_task"
+	CtrlSessionError    = "session_error"
+	CtrlSessionLeave    = "session_leave"
+	CtrlSessionCheckin  = "session_checkin"
 )
 
 const (
