@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v3.21.12
+// 	protoc-gen-go v1.35.2
+// 	protoc        v3.20.3
 // source: services/listenerrpc/service.proto
 
 package listenerrpc
@@ -31,7 +31,7 @@ var file_services_listenerrpc_service_proto_rawDesc = []byte{
 	0x70, 0x62, 0x2f, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a,
 	0x1e, 0x69, 0x6d, 0x70, 0x6c, 0x61, 0x6e, 0x74, 0x2f, 0x69, 0x6d, 0x70, 0x6c, 0x61, 0x6e, 0x74,
 	0x70, 0x62, 0x2f, 0x6d, 0x6f, 0x64, 0x75, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x32,
-	0xe3, 0x0e, 0x0a, 0x0b, 0x4c, 0x69, 0x73, 0x74, 0x65, 0x6e, 0x65, 0x72, 0x52, 0x50, 0x43, 0x12,
+	0x96, 0x0a, 0x0a, 0x0b, 0x4c, 0x69, 0x73, 0x74, 0x65, 0x6e, 0x65, 0x72, 0x52, 0x50, 0x43, 0x12,
 	0x42, 0x0a, 0x0b, 0x53, 0x70, 0x69, 0x74, 0x65, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x12, 0x17,
 	0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x53, 0x70, 0x69, 0x74, 0x65, 0x52,
 	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x1a, 0x16, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74,
@@ -74,112 +74,70 @@ var file_services_listenerrpc_service_proto_rawDesc = []byte{
 	0x0a, 0x0d, 0x4c, 0x69, 0x73, 0x74, 0x50, 0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e, 0x65, 0x73, 0x12,
 	0x12, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x4c, 0x69, 0x73, 0x74, 0x65,
 	0x6e, 0x65, 0x72, 0x1a, 0x13, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x50,
-	0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e, 0x65, 0x73, 0x12, 0x36, 0x0a, 0x0f, 0x52, 0x65, 0x67, 0x69,
+	0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e, 0x65, 0x73, 0x12, 0x40, 0x0a, 0x0f, 0x52, 0x65, 0x67, 0x69,
 	0x73, 0x74, 0x65, 0x72, 0x57, 0x65, 0x62, 0x73, 0x69, 0x74, 0x65, 0x12, 0x12, 0x2e, 0x63, 0x6c,
 	0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x50, 0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e, 0x65, 0x1a,
-	0x0f, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79,
-	0x12, 0x37, 0x0a, 0x0c, 0x53, 0x74, 0x61, 0x72, 0x74, 0x57, 0x65, 0x62, 0x73, 0x69, 0x74, 0x65,
-	0x12, 0x16, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x43, 0x74, 0x72, 0x6c,
-	0x50, 0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e, 0x65, 0x1a, 0x0f, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e,
-	0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x12, 0x36, 0x0a, 0x0b, 0x53, 0x74, 0x6f,
-	0x70, 0x57, 0x65, 0x62, 0x73, 0x69, 0x74, 0x65, 0x12, 0x16, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e,
-	0x74, 0x70, 0x62, 0x2e, 0x43, 0x74, 0x72, 0x6c, 0x50, 0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e, 0x65,
-	0x1a, 0x0f, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74,
-	0x79, 0x12, 0x38, 0x0a, 0x0d, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x57, 0x65, 0x62, 0x73, 0x69,
+	0x19, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x57, 0x65, 0x62, 0x73, 0x69,
+	0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x37, 0x0a, 0x0c, 0x53, 0x74,
+	0x61, 0x72, 0x74, 0x57, 0x65, 0x62, 0x73, 0x69, 0x74, 0x65, 0x12, 0x16, 0x2e, 0x63, 0x6c, 0x69,
+	0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x43, 0x74, 0x72, 0x6c, 0x50, 0x69, 0x70, 0x65, 0x6c, 0x69,
+	0x6e, 0x65, 0x1a, 0x0f, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d,
+	0x70, 0x74, 0x79, 0x12, 0x36, 0x0a, 0x0b, 0x53, 0x74, 0x6f, 0x70, 0x57, 0x65, 0x62, 0x73, 0x69,
 	0x74, 0x65, 0x12, 0x16, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x43, 0x74,
 	0x72, 0x6c, 0x50, 0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e, 0x65, 0x1a, 0x0f, 0x2e, 0x63, 0x6c, 0x69,
-	0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x12, 0x37, 0x0a, 0x0c, 0x4c,
-	0x69, 0x73, 0x74, 0x57, 0x65, 0x62, 0x73, 0x69, 0x74, 0x65, 0x73, 0x12, 0x12, 0x2e, 0x63, 0x6c,
-	0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x4c, 0x69, 0x73, 0x74, 0x65, 0x6e, 0x65, 0x72, 0x1a,
-	0x13, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x50, 0x69, 0x70, 0x65, 0x6c,
-	0x69, 0x6e, 0x65, 0x73, 0x12, 0x3a, 0x0a, 0x0e, 0x4c, 0x69, 0x73, 0x74, 0x57, 0x65, 0x62, 0x43,
-	0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x12, 0x11, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70,
-	0x62, 0x2e, 0x57, 0x65, 0x62, 0x73, 0x69, 0x74, 0x65, 0x1a, 0x15, 0x2e, 0x63, 0x6c, 0x69, 0x65,
-	0x6e, 0x74, 0x70, 0x62, 0x2e, 0x57, 0x65, 0x62, 0x43, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x73,
-	0x12, 0x37, 0x0a, 0x11, 0x57, 0x65, 0x62, 0x73, 0x69, 0x74, 0x65, 0x41, 0x64, 0x64, 0x43, 0x6f,
-	0x6e, 0x74, 0x65, 0x6e, 0x74, 0x12, 0x11, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62,
-	0x2e, 0x57, 0x65, 0x62, 0x73, 0x69, 0x74, 0x65, 0x1a, 0x0f, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e,
-	0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x12, 0x3d, 0x0a, 0x14, 0x57, 0x65, 0x62,
-	0x73, 0x69, 0x74, 0x65, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x43, 0x6f, 0x6e, 0x74, 0x65, 0x6e,
-	0x74, 0x12, 0x14, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x57, 0x65, 0x62,
-	0x43, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x1a, 0x0f, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74,
-	0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x12, 0x3d, 0x0a, 0x14, 0x57, 0x65, 0x62, 0x73,
-	0x69, 0x74, 0x65, 0x52, 0x65, 0x6d, 0x6f, 0x76, 0x65, 0x43, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74,
-	0x12, 0x14, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x57, 0x65, 0x62, 0x43,
-	0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x1a, 0x0f, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70,
-	0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x12, 0x32, 0x0a, 0x0b, 0x52, 0x65, 0x67, 0x69, 0x73,
-	0x74, 0x65, 0x72, 0x52, 0x65, 0x6d, 0x12, 0x12, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70,
-	0x62, 0x2e, 0x50, 0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e, 0x65, 0x1a, 0x0f, 0x2e, 0x63, 0x6c, 0x69,
-	0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x12, 0x33, 0x0a, 0x08, 0x53,
-	0x74, 0x61, 0x72, 0x74, 0x52, 0x65, 0x6d, 0x12, 0x16, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74,
-	0x70, 0x62, 0x2e, 0x43, 0x74, 0x72, 0x6c, 0x50, 0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e, 0x65, 0x1a,
+	0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x12, 0x33, 0x0a, 0x0d, 0x55,
+	0x70, 0x6c, 0x6f, 0x61, 0x64, 0x57, 0x65, 0x62, 0x73, 0x69, 0x74, 0x65, 0x12, 0x11, 0x2e, 0x63,
+	0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x57, 0x65, 0x62, 0x73, 0x69, 0x74, 0x65, 0x1a,
 	0x0f, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79,
-	0x12, 0x32, 0x0a, 0x07, 0x53, 0x74, 0x6f, 0x70, 0x52, 0x65, 0x6d, 0x12, 0x16, 0x2e, 0x63, 0x6c,
-	0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x43, 0x74, 0x72, 0x6c, 0x50, 0x69, 0x70, 0x65, 0x6c,
-	0x69, 0x6e, 0x65, 0x1a, 0x0f, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45,
-	0x6d, 0x70, 0x74, 0x79, 0x12, 0x34, 0x0a, 0x09, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x52, 0x65,
-	0x6d, 0x12, 0x16, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x43, 0x74, 0x72,
-	0x6c, 0x50, 0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e, 0x65, 0x1a, 0x0f, 0x2e, 0x63, 0x6c, 0x69, 0x65,
-	0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x12, 0x33, 0x0a, 0x08, 0x4c, 0x69,
-	0x73, 0x74, 0x52, 0x65, 0x6d, 0x73, 0x12, 0x12, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70,
-	0x62, 0x2e, 0x4c, 0x69, 0x73, 0x74, 0x65, 0x6e, 0x65, 0x72, 0x1a, 0x13, 0x2e, 0x63, 0x6c, 0x69,
-	0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x50, 0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e, 0x65, 0x73, 0x12,
-	0x30, 0x0a, 0x0a, 0x4e, 0x65, 0x77, 0x50, 0x72, 0x6f, 0x66, 0x69, 0x6c, 0x65, 0x12, 0x11, 0x2e,
-	0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x50, 0x72, 0x6f, 0x66, 0x69, 0x6c, 0x65,
-	0x1a, 0x0f, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74,
-	0x79, 0x12, 0x36, 0x0a, 0x0c, 0x46, 0x69, 0x6e, 0x64, 0x41, 0x72, 0x74, 0x69, 0x66, 0x61, 0x63,
-	0x74, 0x12, 0x12, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x41, 0x72, 0x74,
-	0x69, 0x66, 0x61, 0x63, 0x74, 0x1a, 0x12, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62,
-	0x2e, 0x41, 0x72, 0x74, 0x69, 0x66, 0x61, 0x63, 0x74, 0x12, 0x2e, 0x0a, 0x05, 0x42, 0x75, 0x69,
-	0x6c, 0x64, 0x12, 0x12, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x47, 0x65,
-	0x6e, 0x65, 0x72, 0x61, 0x74, 0x65, 0x1a, 0x11, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70,
-	0x62, 0x2e, 0x42, 0x75, 0x69, 0x6c, 0x64, 0x65, 0x72, 0x12, 0x30, 0x0a, 0x0c, 0x44, 0x6f, 0x63,
-	0x6b, 0x65, 0x72, 0x53, 0x74, 0x61, 0x74, 0x75, 0x73, 0x12, 0x0f, 0x2e, 0x63, 0x6c, 0x69, 0x65,
-	0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x1a, 0x0f, 0x2e, 0x63, 0x6c, 0x69,
-	0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x12, 0x4d, 0x0a, 0x17, 0x54,
-	0x72, 0x69, 0x67, 0x67, 0x65, 0x72, 0x57, 0x6f, 0x72, 0x6b, 0x66, 0x6c, 0x6f, 0x77, 0x44, 0x69,
-	0x73, 0x70, 0x61, 0x74, 0x63, 0x68, 0x12, 0x1f, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70,
-	0x62, 0x2e, 0x47, 0x69, 0x74, 0x68, 0x75, 0x62, 0x57, 0x6f, 0x72, 0x6b, 0x66, 0x6c, 0x6f, 0x77,
-	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x11, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74,
-	0x70, 0x62, 0x2e, 0x42, 0x75, 0x69, 0x6c, 0x64, 0x65, 0x72, 0x12, 0x42, 0x0a, 0x0e, 0x57, 0x6f,
-	0x72, 0x6b, 0x66, 0x6c, 0x6f, 0x77, 0x53, 0x74, 0x61, 0x74, 0x75, 0x73, 0x12, 0x1f, 0x2e, 0x63,
-	0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x47, 0x69, 0x74, 0x68, 0x75, 0x62, 0x57, 0x6f,
-	0x72, 0x6b, 0x66, 0x6c, 0x6f, 0x77, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x0f, 0x2e,
-	0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x12, 0x35,
-	0x0a, 0x0b, 0x47, 0x65, 0x74, 0x41, 0x72, 0x74, 0x69, 0x66, 0x61, 0x63, 0x74, 0x12, 0x12, 0x2e,
-	0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x41, 0x72, 0x74, 0x69, 0x66, 0x61, 0x63,
-	0x74, 0x1a, 0x12, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x41, 0x72, 0x74,
-	0x69, 0x66, 0x61, 0x63, 0x74, 0x42, 0x4b, 0x5a, 0x49, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e,
-	0x63, 0x6f, 0x6d, 0x2f, 0x63, 0x68, 0x61, 0x69, 0x6e, 0x72, 0x65, 0x61, 0x63, 0x74, 0x6f, 0x72,
-	0x73, 0x2f, 0x6d, 0x61, 0x6c, 0x69, 0x63, 0x65, 0x2d, 0x6e, 0x65, 0x74, 0x77, 0x6f, 0x72, 0x6b,
-	0x2f, 0x68, 0x65, 0x6c, 0x70, 0x65, 0x72, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x73, 0x65,
-	0x72, 0x76, 0x69, 0x63, 0x65, 0x73, 0x2f, 0x6c, 0x69, 0x73, 0x74, 0x65, 0x6e, 0x65, 0x72, 0x72,
-	0x70, 0x63, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x12, 0x37, 0x0a, 0x0c, 0x4c, 0x69, 0x73, 0x74, 0x57, 0x65, 0x62, 0x73, 0x69, 0x74, 0x65, 0x73,
+	0x12, 0x12, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x4c, 0x69, 0x73, 0x74,
+	0x65, 0x6e, 0x65, 0x72, 0x1a, 0x13, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e,
+	0x50, 0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e, 0x65, 0x73, 0x12, 0x32, 0x0a, 0x0b, 0x52, 0x65, 0x67,
+	0x69, 0x73, 0x74, 0x65, 0x72, 0x52, 0x65, 0x6d, 0x12, 0x12, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e,
+	0x74, 0x70, 0x62, 0x2e, 0x50, 0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e, 0x65, 0x1a, 0x0f, 0x2e, 0x63,
+	0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x12, 0x33, 0x0a,
+	0x08, 0x53, 0x74, 0x61, 0x72, 0x74, 0x52, 0x65, 0x6d, 0x12, 0x16, 0x2e, 0x63, 0x6c, 0x69, 0x65,
+	0x6e, 0x74, 0x70, 0x62, 0x2e, 0x43, 0x74, 0x72, 0x6c, 0x50, 0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e,
+	0x65, 0x1a, 0x0f, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70,
+	0x74, 0x79, 0x12, 0x32, 0x0a, 0x07, 0x53, 0x74, 0x6f, 0x70, 0x52, 0x65, 0x6d, 0x12, 0x16, 0x2e,
+	0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x43, 0x74, 0x72, 0x6c, 0x50, 0x69, 0x70,
+	0x65, 0x6c, 0x69, 0x6e, 0x65, 0x1a, 0x0f, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62,
+	0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x12, 0x34, 0x0a, 0x09, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65,
+	0x52, 0x65, 0x6d, 0x12, 0x16, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x43,
+	0x74, 0x72, 0x6c, 0x50, 0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e, 0x65, 0x1a, 0x0f, 0x2e, 0x63, 0x6c,
+	0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x12, 0x33, 0x0a, 0x08,
+	0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x6d, 0x73, 0x12, 0x12, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e,
+	0x74, 0x70, 0x62, 0x2e, 0x4c, 0x69, 0x73, 0x74, 0x65, 0x6e, 0x65, 0x72, 0x1a, 0x13, 0x2e, 0x63,
+	0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x50, 0x69, 0x70, 0x65, 0x6c, 0x69, 0x6e, 0x65,
+	0x73, 0x12, 0x35, 0x0a, 0x0b, 0x47, 0x65, 0x74, 0x41, 0x72, 0x74, 0x69, 0x66, 0x61, 0x63, 0x74,
+	0x12, 0x12, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e, 0x41, 0x72, 0x74, 0x69,
+	0x66, 0x61, 0x63, 0x74, 0x1a, 0x12, 0x2e, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x70, 0x62, 0x2e,
+	0x41, 0x72, 0x74, 0x69, 0x66, 0x61, 0x63, 0x74, 0x42, 0x4b, 0x5a, 0x49, 0x67, 0x69, 0x74, 0x68,
+	0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x63, 0x68, 0x61, 0x69, 0x6e, 0x72, 0x65, 0x61, 0x63,
+	0x74, 0x6f, 0x72, 0x73, 0x2f, 0x6d, 0x61, 0x6c, 0x69, 0x63, 0x65, 0x2d, 0x6e, 0x65, 0x74, 0x77,
+	0x6f, 0x72, 0x6b, 0x2f, 0x68, 0x65, 0x6c, 0x70, 0x65, 0x72, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f,
+	0x2f, 0x73, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x73, 0x2f, 0x6c, 0x69, 0x73, 0x74, 0x65, 0x6e,
+	0x65, 0x72, 0x72, 0x70, 0x63, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
-var file_services_listenerrpc_service_proto_goTypes = []interface{}{
-	(*clientpb.SpiteResponse)(nil),         // 0: clientpb.SpiteResponse
-	(*clientpb.JobStatus)(nil),             // 1: clientpb.JobStatus
-	(*clientpb.RegisterSession)(nil),       // 2: clientpb.RegisterSession
-	(*implantpb.SysInfo)(nil),              // 3: modulepb.SysInfo
-	(*implantpb.Ping)(nil),                 // 4: modulepb.Ping
-	(*implantpb.Request)(nil),              // 5: modulepb.Request
-	(*clientpb.RegisterListener)(nil),      // 6: clientpb.RegisterListener
-	(*clientpb.Pipeline)(nil),              // 7: clientpb.Pipeline
-	(*clientpb.CtrlPipeline)(nil),          // 8: clientpb.CtrlPipeline
-	(*clientpb.Listener)(nil),              // 9: clientpb.Listener
-	(*clientpb.Website)(nil),               // 10: clientpb.Website
-	(*clientpb.WebContent)(nil),            // 11: clientpb.WebContent
-	(*clientpb.Profile)(nil),               // 12: clientpb.Profile
-	(*clientpb.Artifact)(nil),              // 13: clientpb.Artifact
-	(*clientpb.Generate)(nil),              // 14: clientpb.Generate
-	(*clientpb.Empty)(nil),                 // 15: clientpb.Empty
-	(*clientpb.GithubWorkflowRequest)(nil), // 16: clientpb.GithubWorkflowRequest
-	(*clientpb.SpiteRequest)(nil),          // 17: clientpb.SpiteRequest
-	(*clientpb.JobCtrl)(nil),               // 18: clientpb.JobCtrl
-	(*clientpb.Pipelines)(nil),             // 19: clientpb.Pipelines
-	(*clientpb.WebContents)(nil),           // 20: clientpb.WebContents
-	(*clientpb.Builder)(nil),               // 21: clientpb.Builder
+var file_services_listenerrpc_service_proto_goTypes = []any{
+	(*clientpb.SpiteResponse)(nil),    // 0: clientpb.SpiteResponse
+	(*clientpb.JobStatus)(nil),        // 1: clientpb.JobStatus
+	(*clientpb.RegisterSession)(nil),  // 2: clientpb.RegisterSession
+	(*implantpb.SysInfo)(nil),         // 3: modulepb.SysInfo
+	(*implantpb.Ping)(nil),            // 4: modulepb.Ping
+	(*implantpb.Request)(nil),         // 5: modulepb.Request
+	(*clientpb.RegisterListener)(nil), // 6: clientpb.RegisterListener
+	(*clientpb.Pipeline)(nil),         // 7: clientpb.Pipeline
+	(*clientpb.CtrlPipeline)(nil),     // 8: clientpb.CtrlPipeline
+	(*clientpb.Listener)(nil),         // 9: clientpb.Listener
+	(*clientpb.Website)(nil),          // 10: clientpb.Website
+	(*clientpb.Artifact)(nil),         // 11: clientpb.Artifact
+	(*clientpb.SpiteRequest)(nil),     // 12: clientpb.SpiteRequest
+	(*clientpb.JobCtrl)(nil),          // 13: clientpb.JobCtrl
+	(*clientpb.Empty)(nil),            // 14: clientpb.Empty
+	(*clientpb.Pipelines)(nil),        // 15: clientpb.Pipelines
+	(*clientpb.WebsiteResponse)(nil),  // 16: clientpb.WebsiteResponse
 }
 var file_services_listenerrpc_service_proto_depIdxs = []int32{
 	0,  // 0: listenerrpc.ListenerRPC.SpiteStream:input_type -> clientpb.SpiteResponse
@@ -197,59 +155,39 @@ var file_services_listenerrpc_service_proto_depIdxs = []int32{
 	7,  // 12: listenerrpc.ListenerRPC.RegisterWebsite:input_type -> clientpb.Pipeline
 	8,  // 13: listenerrpc.ListenerRPC.StartWebsite:input_type -> clientpb.CtrlPipeline
 	8,  // 14: listenerrpc.ListenerRPC.StopWebsite:input_type -> clientpb.CtrlPipeline
-	8,  // 15: listenerrpc.ListenerRPC.DeleteWebsite:input_type -> clientpb.CtrlPipeline
+	10, // 15: listenerrpc.ListenerRPC.UploadWebsite:input_type -> clientpb.Website
 	9,  // 16: listenerrpc.ListenerRPC.ListWebsites:input_type -> clientpb.Listener
-	10, // 17: listenerrpc.ListenerRPC.ListWebContent:input_type -> clientpb.Website
-	10, // 18: listenerrpc.ListenerRPC.WebsiteAddContent:input_type -> clientpb.Website
-	11, // 19: listenerrpc.ListenerRPC.WebsiteUpdateContent:input_type -> clientpb.WebContent
-	11, // 20: listenerrpc.ListenerRPC.WebsiteRemoveContent:input_type -> clientpb.WebContent
-	7,  // 21: listenerrpc.ListenerRPC.RegisterRem:input_type -> clientpb.Pipeline
-	8,  // 22: listenerrpc.ListenerRPC.StartRem:input_type -> clientpb.CtrlPipeline
-	8,  // 23: listenerrpc.ListenerRPC.StopRem:input_type -> clientpb.CtrlPipeline
-	8,  // 24: listenerrpc.ListenerRPC.DeleteRem:input_type -> clientpb.CtrlPipeline
-	9,  // 25: listenerrpc.ListenerRPC.ListRems:input_type -> clientpb.Listener
-	12, // 26: listenerrpc.ListenerRPC.NewProfile:input_type -> clientpb.Profile
-	13, // 27: listenerrpc.ListenerRPC.FindArtifact:input_type -> clientpb.Artifact
-	14, // 28: listenerrpc.ListenerRPC.Build:input_type -> clientpb.Generate
-	15, // 29: listenerrpc.ListenerRPC.DockerStatus:input_type -> clientpb.Empty
-	16, // 30: listenerrpc.ListenerRPC.TriggerWorkflowDispatch:input_type -> clientpb.GithubWorkflowRequest
-	16, // 31: listenerrpc.ListenerRPC.WorkflowStatus:input_type -> clientpb.GithubWorkflowRequest
-	13, // 32: listenerrpc.ListenerRPC.GetArtifact:input_type -> clientpb.Artifact
-	17, // 33: listenerrpc.ListenerRPC.SpiteStream:output_type -> clientpb.SpiteRequest
-	18, // 34: listenerrpc.ListenerRPC.JobStream:output_type -> clientpb.JobCtrl
-	15, // 35: listenerrpc.ListenerRPC.Register:output_type -> clientpb.Empty
-	15, // 36: listenerrpc.ListenerRPC.SysInfo:output_type -> clientpb.Empty
-	15, // 37: listenerrpc.ListenerRPC.Checkin:output_type -> clientpb.Empty
-	15, // 38: listenerrpc.ListenerRPC.InitBindSession:output_type -> clientpb.Empty
-	15, // 39: listenerrpc.ListenerRPC.RegisterListener:output_type -> clientpb.Empty
-	15, // 40: listenerrpc.ListenerRPC.RegisterPipeline:output_type -> clientpb.Empty
-	15, // 41: listenerrpc.ListenerRPC.StartPipeline:output_type -> clientpb.Empty
-	15, // 42: listenerrpc.ListenerRPC.StopPipeline:output_type -> clientpb.Empty
-	15, // 43: listenerrpc.ListenerRPC.DeletePipeline:output_type -> clientpb.Empty
-	19, // 44: listenerrpc.ListenerRPC.ListPipelines:output_type -> clientpb.Pipelines
-	15, // 45: listenerrpc.ListenerRPC.RegisterWebsite:output_type -> clientpb.Empty
-	15, // 46: listenerrpc.ListenerRPC.StartWebsite:output_type -> clientpb.Empty
-	15, // 47: listenerrpc.ListenerRPC.StopWebsite:output_type -> clientpb.Empty
-	15, // 48: listenerrpc.ListenerRPC.DeleteWebsite:output_type -> clientpb.Empty
-	19, // 49: listenerrpc.ListenerRPC.ListWebsites:output_type -> clientpb.Pipelines
-	20, // 50: listenerrpc.ListenerRPC.ListWebContent:output_type -> clientpb.WebContents
-	15, // 51: listenerrpc.ListenerRPC.WebsiteAddContent:output_type -> clientpb.Empty
-	15, // 52: listenerrpc.ListenerRPC.WebsiteUpdateContent:output_type -> clientpb.Empty
-	15, // 53: listenerrpc.ListenerRPC.WebsiteRemoveContent:output_type -> clientpb.Empty
-	15, // 54: listenerrpc.ListenerRPC.RegisterRem:output_type -> clientpb.Empty
-	15, // 55: listenerrpc.ListenerRPC.StartRem:output_type -> clientpb.Empty
-	15, // 56: listenerrpc.ListenerRPC.StopRem:output_type -> clientpb.Empty
-	15, // 57: listenerrpc.ListenerRPC.DeleteRem:output_type -> clientpb.Empty
-	19, // 58: listenerrpc.ListenerRPC.ListRems:output_type -> clientpb.Pipelines
-	15, // 59: listenerrpc.ListenerRPC.NewProfile:output_type -> clientpb.Empty
-	13, // 60: listenerrpc.ListenerRPC.FindArtifact:output_type -> clientpb.Artifact
-	21, // 61: listenerrpc.ListenerRPC.Build:output_type -> clientpb.Builder
-	15, // 62: listenerrpc.ListenerRPC.DockerStatus:output_type -> clientpb.Empty
-	21, // 63: listenerrpc.ListenerRPC.TriggerWorkflowDispatch:output_type -> clientpb.Builder
-	15, // 64: listenerrpc.ListenerRPC.WorkflowStatus:output_type -> clientpb.Empty
-	13, // 65: listenerrpc.ListenerRPC.GetArtifact:output_type -> clientpb.Artifact
-	33, // [33:66] is the sub-list for method output_type
-	0,  // [0:33] is the sub-list for method input_type
+	7,  // 17: listenerrpc.ListenerRPC.RegisterRem:input_type -> clientpb.Pipeline
+	8,  // 18: listenerrpc.ListenerRPC.StartRem:input_type -> clientpb.CtrlPipeline
+	8,  // 19: listenerrpc.ListenerRPC.StopRem:input_type -> clientpb.CtrlPipeline
+	8,  // 20: listenerrpc.ListenerRPC.DeleteRem:input_type -> clientpb.CtrlPipeline
+	9,  // 21: listenerrpc.ListenerRPC.ListRems:input_type -> clientpb.Listener
+	11, // 22: listenerrpc.ListenerRPC.GetArtifact:input_type -> clientpb.Artifact
+	12, // 23: listenerrpc.ListenerRPC.SpiteStream:output_type -> clientpb.SpiteRequest
+	13, // 24: listenerrpc.ListenerRPC.JobStream:output_type -> clientpb.JobCtrl
+	14, // 25: listenerrpc.ListenerRPC.Register:output_type -> clientpb.Empty
+	14, // 26: listenerrpc.ListenerRPC.SysInfo:output_type -> clientpb.Empty
+	14, // 27: listenerrpc.ListenerRPC.Checkin:output_type -> clientpb.Empty
+	14, // 28: listenerrpc.ListenerRPC.InitBindSession:output_type -> clientpb.Empty
+	14, // 29: listenerrpc.ListenerRPC.RegisterListener:output_type -> clientpb.Empty
+	14, // 30: listenerrpc.ListenerRPC.RegisterPipeline:output_type -> clientpb.Empty
+	14, // 31: listenerrpc.ListenerRPC.StartPipeline:output_type -> clientpb.Empty
+	14, // 32: listenerrpc.ListenerRPC.StopPipeline:output_type -> clientpb.Empty
+	14, // 33: listenerrpc.ListenerRPC.DeletePipeline:output_type -> clientpb.Empty
+	15, // 34: listenerrpc.ListenerRPC.ListPipelines:output_type -> clientpb.Pipelines
+	16, // 35: listenerrpc.ListenerRPC.RegisterWebsite:output_type -> clientpb.WebsiteResponse
+	14, // 36: listenerrpc.ListenerRPC.StartWebsite:output_type -> clientpb.Empty
+	14, // 37: listenerrpc.ListenerRPC.StopWebsite:output_type -> clientpb.Empty
+	14, // 38: listenerrpc.ListenerRPC.UploadWebsite:output_type -> clientpb.Empty
+	15, // 39: listenerrpc.ListenerRPC.ListWebsites:output_type -> clientpb.Pipelines
+	14, // 40: listenerrpc.ListenerRPC.RegisterRem:output_type -> clientpb.Empty
+	14, // 41: listenerrpc.ListenerRPC.StartRem:output_type -> clientpb.Empty
+	14, // 42: listenerrpc.ListenerRPC.StopRem:output_type -> clientpb.Empty
+	14, // 43: listenerrpc.ListenerRPC.DeleteRem:output_type -> clientpb.Empty
+	15, // 44: listenerrpc.ListenerRPC.ListRems:output_type -> clientpb.Pipelines
+	11, // 45: listenerrpc.ListenerRPC.GetArtifact:output_type -> clientpb.Artifact
+	23, // [23:46] is the sub-list for method output_type
+	0,  // [0:23] is the sub-list for method input_type
 	0,  // [0:0] is the sub-list for extension type_name
 	0,  // [0:0] is the sub-list for extension extendee
 	0,  // [0:0] is the sub-list for field type_name
