@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v3.21.12
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v3.20.3
 // source: services/listenerrpc/service.proto
 
 package listenerrpc
@@ -17,51 +17,41 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
-	ListenerRPC_SpiteStream_FullMethodName             = "/listenerrpc.ListenerRPC/SpiteStream"
-	ListenerRPC_JobStream_FullMethodName               = "/listenerrpc.ListenerRPC/JobStream"
-	ListenerRPC_Register_FullMethodName                = "/listenerrpc.ListenerRPC/Register"
-	ListenerRPC_SysInfo_FullMethodName                 = "/listenerrpc.ListenerRPC/SysInfo"
-	ListenerRPC_Checkin_FullMethodName                 = "/listenerrpc.ListenerRPC/Checkin"
-	ListenerRPC_InitBindSession_FullMethodName         = "/listenerrpc.ListenerRPC/InitBindSession"
-	ListenerRPC_RegisterListener_FullMethodName        = "/listenerrpc.ListenerRPC/RegisterListener"
-	ListenerRPC_RegisterPipeline_FullMethodName        = "/listenerrpc.ListenerRPC/RegisterPipeline"
-	ListenerRPC_StartPipeline_FullMethodName           = "/listenerrpc.ListenerRPC/StartPipeline"
-	ListenerRPC_StopPipeline_FullMethodName            = "/listenerrpc.ListenerRPC/StopPipeline"
-	ListenerRPC_DeletePipeline_FullMethodName          = "/listenerrpc.ListenerRPC/DeletePipeline"
-	ListenerRPC_ListPipelines_FullMethodName           = "/listenerrpc.ListenerRPC/ListPipelines"
-	ListenerRPC_RegisterWebsite_FullMethodName         = "/listenerrpc.ListenerRPC/RegisterWebsite"
-	ListenerRPC_StartWebsite_FullMethodName            = "/listenerrpc.ListenerRPC/StartWebsite"
-	ListenerRPC_StopWebsite_FullMethodName             = "/listenerrpc.ListenerRPC/StopWebsite"
-	ListenerRPC_DeleteWebsite_FullMethodName           = "/listenerrpc.ListenerRPC/DeleteWebsite"
-	ListenerRPC_ListWebsites_FullMethodName            = "/listenerrpc.ListenerRPC/ListWebsites"
-	ListenerRPC_ListWebContent_FullMethodName          = "/listenerrpc.ListenerRPC/ListWebContent"
-	ListenerRPC_WebsiteAddContent_FullMethodName       = "/listenerrpc.ListenerRPC/WebsiteAddContent"
-	ListenerRPC_WebsiteUpdateContent_FullMethodName    = "/listenerrpc.ListenerRPC/WebsiteUpdateContent"
-	ListenerRPC_WebsiteRemoveContent_FullMethodName    = "/listenerrpc.ListenerRPC/WebsiteRemoveContent"
-	ListenerRPC_RegisterRem_FullMethodName             = "/listenerrpc.ListenerRPC/RegisterRem"
-	ListenerRPC_StartRem_FullMethodName                = "/listenerrpc.ListenerRPC/StartRem"
-	ListenerRPC_StopRem_FullMethodName                 = "/listenerrpc.ListenerRPC/StopRem"
-	ListenerRPC_DeleteRem_FullMethodName               = "/listenerrpc.ListenerRPC/DeleteRem"
-	ListenerRPC_ListRems_FullMethodName                = "/listenerrpc.ListenerRPC/ListRems"
-	ListenerRPC_NewProfile_FullMethodName              = "/listenerrpc.ListenerRPC/NewProfile"
-	ListenerRPC_FindArtifact_FullMethodName            = "/listenerrpc.ListenerRPC/FindArtifact"
-	ListenerRPC_Build_FullMethodName                   = "/listenerrpc.ListenerRPC/Build"
-	ListenerRPC_DockerStatus_FullMethodName            = "/listenerrpc.ListenerRPC/DockerStatus"
-	ListenerRPC_TriggerWorkflowDispatch_FullMethodName = "/listenerrpc.ListenerRPC/TriggerWorkflowDispatch"
-	ListenerRPC_WorkflowStatus_FullMethodName          = "/listenerrpc.ListenerRPC/WorkflowStatus"
-	ListenerRPC_GetArtifact_FullMethodName             = "/listenerrpc.ListenerRPC/GetArtifact"
+	ListenerRPC_SpiteStream_FullMethodName      = "/listenerrpc.ListenerRPC/SpiteStream"
+	ListenerRPC_JobStream_FullMethodName        = "/listenerrpc.ListenerRPC/JobStream"
+	ListenerRPC_Register_FullMethodName         = "/listenerrpc.ListenerRPC/Register"
+	ListenerRPC_SysInfo_FullMethodName          = "/listenerrpc.ListenerRPC/SysInfo"
+	ListenerRPC_Checkin_FullMethodName          = "/listenerrpc.ListenerRPC/Checkin"
+	ListenerRPC_InitBindSession_FullMethodName  = "/listenerrpc.ListenerRPC/InitBindSession"
+	ListenerRPC_RegisterListener_FullMethodName = "/listenerrpc.ListenerRPC/RegisterListener"
+	ListenerRPC_RegisterPipeline_FullMethodName = "/listenerrpc.ListenerRPC/RegisterPipeline"
+	ListenerRPC_StartPipeline_FullMethodName    = "/listenerrpc.ListenerRPC/StartPipeline"
+	ListenerRPC_StopPipeline_FullMethodName     = "/listenerrpc.ListenerRPC/StopPipeline"
+	ListenerRPC_DeletePipeline_FullMethodName   = "/listenerrpc.ListenerRPC/DeletePipeline"
+	ListenerRPC_ListPipelines_FullMethodName    = "/listenerrpc.ListenerRPC/ListPipelines"
+	ListenerRPC_RegisterWebsite_FullMethodName  = "/listenerrpc.ListenerRPC/RegisterWebsite"
+	ListenerRPC_StartWebsite_FullMethodName     = "/listenerrpc.ListenerRPC/StartWebsite"
+	ListenerRPC_StopWebsite_FullMethodName      = "/listenerrpc.ListenerRPC/StopWebsite"
+	ListenerRPC_UploadWebsite_FullMethodName    = "/listenerrpc.ListenerRPC/UploadWebsite"
+	ListenerRPC_ListWebsites_FullMethodName     = "/listenerrpc.ListenerRPC/ListWebsites"
+	ListenerRPC_RegisterRem_FullMethodName      = "/listenerrpc.ListenerRPC/RegisterRem"
+	ListenerRPC_StartRem_FullMethodName         = "/listenerrpc.ListenerRPC/StartRem"
+	ListenerRPC_StopRem_FullMethodName          = "/listenerrpc.ListenerRPC/StopRem"
+	ListenerRPC_DeleteRem_FullMethodName        = "/listenerrpc.ListenerRPC/DeleteRem"
+	ListenerRPC_ListRems_FullMethodName         = "/listenerrpc.ListenerRPC/ListRems"
+	ListenerRPC_GetArtifact_FullMethodName      = "/listenerrpc.ListenerRPC/GetArtifact"
 )
 
 // ListenerRPCClient is the client API for ListenerRPC service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
 type ListenerRPCClient interface {
-	SpiteStream(ctx context.Context, opts ...grpc.CallOption) (ListenerRPC_SpiteStreamClient, error)
-	JobStream(ctx context.Context, opts ...grpc.CallOption) (ListenerRPC_JobStreamClient, error)
+	SpiteStream(ctx context.Context, opts ...grpc.CallOption) (grpc.BidiStreamingClient[clientpb.SpiteResponse, clientpb.SpiteRequest], error)
+	JobStream(ctx context.Context, opts ...grpc.CallOption) (grpc.BidiStreamingClient[clientpb.JobStatus, clientpb.JobCtrl], error)
 	// implant
 	Register(ctx context.Context, in *clientpb.RegisterSession, opts ...grpc.CallOption) (*clientpb.Empty, error)
 	SysInfo(ctx context.Context, in *implantpb.SysInfo, opts ...grpc.CallOption) (*clientpb.Empty, error)
@@ -75,29 +65,17 @@ type ListenerRPCClient interface {
 	DeletePipeline(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error)
 	ListPipelines(ctx context.Context, in *clientpb.Listener, opts ...grpc.CallOption) (*clientpb.Pipelines, error)
 	// website
-	RegisterWebsite(ctx context.Context, in *clientpb.Pipeline, opts ...grpc.CallOption) (*clientpb.Empty, error)
+	RegisterWebsite(ctx context.Context, in *clientpb.Pipeline, opts ...grpc.CallOption) (*clientpb.WebsiteResponse, error)
 	StartWebsite(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error)
 	StopWebsite(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error)
-	DeleteWebsite(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error)
+	UploadWebsite(ctx context.Context, in *clientpb.Website, opts ...grpc.CallOption) (*clientpb.Empty, error)
 	ListWebsites(ctx context.Context, in *clientpb.Listener, opts ...grpc.CallOption) (*clientpb.Pipelines, error)
-	ListWebContent(ctx context.Context, in *clientpb.Website, opts ...grpc.CallOption) (*clientpb.WebContents, error)
-	WebsiteAddContent(ctx context.Context, in *clientpb.Website, opts ...grpc.CallOption) (*clientpb.Empty, error)
-	WebsiteUpdateContent(ctx context.Context, in *clientpb.WebContent, opts ...grpc.CallOption) (*clientpb.Empty, error)
-	WebsiteRemoveContent(ctx context.Context, in *clientpb.WebContent, opts ...grpc.CallOption) (*clientpb.Empty, error)
 	// rem
 	RegisterRem(ctx context.Context, in *clientpb.Pipeline, opts ...grpc.CallOption) (*clientpb.Empty, error)
 	StartRem(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error)
 	StopRem(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error)
 	DeleteRem(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error)
 	ListRems(ctx context.Context, in *clientpb.Listener, opts ...grpc.CallOption) (*clientpb.Pipelines, error)
-	// generator
-	NewProfile(ctx context.Context, in *clientpb.Profile, opts ...grpc.CallOption) (*clientpb.Empty, error)
-	FindArtifact(ctx context.Context, in *clientpb.Artifact, opts ...grpc.CallOption) (*clientpb.Artifact, error)
-	Build(ctx context.Context, in *clientpb.Generate, opts ...grpc.CallOption) (*clientpb.Builder, error)
-	DockerStatus(ctx context.Context, in *clientpb.Empty, opts ...grpc.CallOption) (*clientpb.Empty, error)
-	// action
-	TriggerWorkflowDispatch(ctx context.Context, in *clientpb.GithubWorkflowRequest, opts ...grpc.CallOption) (*clientpb.Builder, error)
-	WorkflowStatus(ctx context.Context, in *clientpb.GithubWorkflowRequest, opts ...grpc.CallOption) (*clientpb.Empty, error)
 	GetArtifact(ctx context.Context, in *clientpb.Artifact, opts ...grpc.CallOption) (*clientpb.Artifact, error)
 }
 
@@ -109,71 +87,36 @@ func NewListenerRPCClient(cc grpc.ClientConnInterface) ListenerRPCClient {
 	return &listenerRPCClient{cc}
 }
 
-func (c *listenerRPCClient) SpiteStream(ctx context.Context, opts ...grpc.CallOption) (ListenerRPC_SpiteStreamClient, error) {
-	stream, err := c.cc.NewStream(ctx, &ListenerRPC_ServiceDesc.Streams[0], ListenerRPC_SpiteStream_FullMethodName, opts...)
+func (c *listenerRPCClient) SpiteStream(ctx context.Context, opts ...grpc.CallOption) (grpc.BidiStreamingClient[clientpb.SpiteResponse, clientpb.SpiteRequest], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &ListenerRPC_ServiceDesc.Streams[0], ListenerRPC_SpiteStream_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &listenerRPCSpiteStreamClient{stream}
+	x := &grpc.GenericClientStream[clientpb.SpiteResponse, clientpb.SpiteRequest]{ClientStream: stream}
 	return x, nil
 }
 
-type ListenerRPC_SpiteStreamClient interface {
-	Send(*clientpb.SpiteResponse) error
-	Recv() (*clientpb.SpiteRequest, error)
-	grpc.ClientStream
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ListenerRPC_SpiteStreamClient = grpc.BidiStreamingClient[clientpb.SpiteResponse, clientpb.SpiteRequest]
 
-type listenerRPCSpiteStreamClient struct {
-	grpc.ClientStream
-}
-
-func (x *listenerRPCSpiteStreamClient) Send(m *clientpb.SpiteResponse) error {
-	return x.ClientStream.SendMsg(m)
-}
-
-func (x *listenerRPCSpiteStreamClient) Recv() (*clientpb.SpiteRequest, error) {
-	m := new(clientpb.SpiteRequest)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-func (c *listenerRPCClient) JobStream(ctx context.Context, opts ...grpc.CallOption) (ListenerRPC_JobStreamClient, error) {
-	stream, err := c.cc.NewStream(ctx, &ListenerRPC_ServiceDesc.Streams[1], ListenerRPC_JobStream_FullMethodName, opts...)
+func (c *listenerRPCClient) JobStream(ctx context.Context, opts ...grpc.CallOption) (grpc.BidiStreamingClient[clientpb.JobStatus, clientpb.JobCtrl], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &ListenerRPC_ServiceDesc.Streams[1], ListenerRPC_JobStream_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &listenerRPCJobStreamClient{stream}
+	x := &grpc.GenericClientStream[clientpb.JobStatus, clientpb.JobCtrl]{ClientStream: stream}
 	return x, nil
 }
 
-type ListenerRPC_JobStreamClient interface {
-	Send(*clientpb.JobStatus) error
-	Recv() (*clientpb.JobCtrl, error)
-	grpc.ClientStream
-}
-
-type listenerRPCJobStreamClient struct {
-	grpc.ClientStream
-}
-
-func (x *listenerRPCJobStreamClient) Send(m *clientpb.JobStatus) error {
-	return x.ClientStream.SendMsg(m)
-}
-
-func (x *listenerRPCJobStreamClient) Recv() (*clientpb.JobCtrl, error) {
-	m := new(clientpb.JobCtrl)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ListenerRPC_JobStreamClient = grpc.BidiStreamingClient[clientpb.JobStatus, clientpb.JobCtrl]
 
 func (c *listenerRPCClient) Register(ctx context.Context, in *clientpb.RegisterSession, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_Register_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_Register_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -181,8 +124,9 @@ func (c *listenerRPCClient) Register(ctx context.Context, in *clientpb.RegisterS
 }
 
 func (c *listenerRPCClient) SysInfo(ctx context.Context, in *implantpb.SysInfo, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_SysInfo_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_SysInfo_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -190,8 +134,9 @@ func (c *listenerRPCClient) SysInfo(ctx context.Context, in *implantpb.SysInfo, 
 }
 
 func (c *listenerRPCClient) Checkin(ctx context.Context, in *implantpb.Ping, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_Checkin_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_Checkin_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -199,8 +144,9 @@ func (c *listenerRPCClient) Checkin(ctx context.Context, in *implantpb.Ping, opt
 }
 
 func (c *listenerRPCClient) InitBindSession(ctx context.Context, in *implantpb.Request, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_InitBindSession_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_InitBindSession_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -208,8 +154,9 @@ func (c *listenerRPCClient) InitBindSession(ctx context.Context, in *implantpb.R
 }
 
 func (c *listenerRPCClient) RegisterListener(ctx context.Context, in *clientpb.RegisterListener, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_RegisterListener_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_RegisterListener_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -217,8 +164,9 @@ func (c *listenerRPCClient) RegisterListener(ctx context.Context, in *clientpb.R
 }
 
 func (c *listenerRPCClient) RegisterPipeline(ctx context.Context, in *clientpb.Pipeline, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_RegisterPipeline_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_RegisterPipeline_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -226,8 +174,9 @@ func (c *listenerRPCClient) RegisterPipeline(ctx context.Context, in *clientpb.P
 }
 
 func (c *listenerRPCClient) StartPipeline(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_StartPipeline_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_StartPipeline_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -235,8 +184,9 @@ func (c *listenerRPCClient) StartPipeline(ctx context.Context, in *clientpb.Ctrl
 }
 
 func (c *listenerRPCClient) StopPipeline(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_StopPipeline_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_StopPipeline_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -244,8 +194,9 @@ func (c *listenerRPCClient) StopPipeline(ctx context.Context, in *clientpb.CtrlP
 }
 
 func (c *listenerRPCClient) DeletePipeline(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_DeletePipeline_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_DeletePipeline_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -253,17 +204,19 @@ func (c *listenerRPCClient) DeletePipeline(ctx context.Context, in *clientpb.Ctr
 }
 
 func (c *listenerRPCClient) ListPipelines(ctx context.Context, in *clientpb.Listener, opts ...grpc.CallOption) (*clientpb.Pipelines, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Pipelines)
-	err := c.cc.Invoke(ctx, ListenerRPC_ListPipelines_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_ListPipelines_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *listenerRPCClient) RegisterWebsite(ctx context.Context, in *clientpb.Pipeline, opts ...grpc.CallOption) (*clientpb.Empty, error) {
-	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_RegisterWebsite_FullMethodName, in, out, opts...)
+func (c *listenerRPCClient) RegisterWebsite(ctx context.Context, in *clientpb.Pipeline, opts ...grpc.CallOption) (*clientpb.WebsiteResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(clientpb.WebsiteResponse)
+	err := c.cc.Invoke(ctx, ListenerRPC_RegisterWebsite_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -271,8 +224,9 @@ func (c *listenerRPCClient) RegisterWebsite(ctx context.Context, in *clientpb.Pi
 }
 
 func (c *listenerRPCClient) StartWebsite(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_StartWebsite_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_StartWebsite_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -280,17 +234,19 @@ func (c *listenerRPCClient) StartWebsite(ctx context.Context, in *clientpb.CtrlP
 }
 
 func (c *listenerRPCClient) StopWebsite(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_StopWebsite_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_StopWebsite_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *listenerRPCClient) DeleteWebsite(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+func (c *listenerRPCClient) UploadWebsite(ctx context.Context, in *clientpb.Website, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_DeleteWebsite_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_UploadWebsite_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -298,44 +254,9 @@ func (c *listenerRPCClient) DeleteWebsite(ctx context.Context, in *clientpb.Ctrl
 }
 
 func (c *listenerRPCClient) ListWebsites(ctx context.Context, in *clientpb.Listener, opts ...grpc.CallOption) (*clientpb.Pipelines, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Pipelines)
-	err := c.cc.Invoke(ctx, ListenerRPC_ListWebsites_FullMethodName, in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *listenerRPCClient) ListWebContent(ctx context.Context, in *clientpb.Website, opts ...grpc.CallOption) (*clientpb.WebContents, error) {
-	out := new(clientpb.WebContents)
-	err := c.cc.Invoke(ctx, ListenerRPC_ListWebContent_FullMethodName, in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *listenerRPCClient) WebsiteAddContent(ctx context.Context, in *clientpb.Website, opts ...grpc.CallOption) (*clientpb.Empty, error) {
-	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_WebsiteAddContent_FullMethodName, in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *listenerRPCClient) WebsiteUpdateContent(ctx context.Context, in *clientpb.WebContent, opts ...grpc.CallOption) (*clientpb.Empty, error) {
-	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_WebsiteUpdateContent_FullMethodName, in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *listenerRPCClient) WebsiteRemoveContent(ctx context.Context, in *clientpb.WebContent, opts ...grpc.CallOption) (*clientpb.Empty, error) {
-	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_WebsiteRemoveContent_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_ListWebsites_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -343,8 +264,9 @@ func (c *listenerRPCClient) WebsiteRemoveContent(ctx context.Context, in *client
 }
 
 func (c *listenerRPCClient) RegisterRem(ctx context.Context, in *clientpb.Pipeline, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_RegisterRem_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_RegisterRem_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -352,8 +274,9 @@ func (c *listenerRPCClient) RegisterRem(ctx context.Context, in *clientpb.Pipeli
 }
 
 func (c *listenerRPCClient) StartRem(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_StartRem_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_StartRem_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -361,8 +284,9 @@ func (c *listenerRPCClient) StartRem(ctx context.Context, in *clientpb.CtrlPipel
 }
 
 func (c *listenerRPCClient) StopRem(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_StopRem_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_StopRem_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -370,8 +294,9 @@ func (c *listenerRPCClient) StopRem(ctx context.Context, in *clientpb.CtrlPipeli
 }
 
 func (c *listenerRPCClient) DeleteRem(ctx context.Context, in *clientpb.CtrlPipeline, opts ...grpc.CallOption) (*clientpb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_DeleteRem_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_DeleteRem_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -379,62 +304,9 @@ func (c *listenerRPCClient) DeleteRem(ctx context.Context, in *clientpb.CtrlPipe
 }
 
 func (c *listenerRPCClient) ListRems(ctx context.Context, in *clientpb.Listener, opts ...grpc.CallOption) (*clientpb.Pipelines, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Pipelines)
-	err := c.cc.Invoke(ctx, ListenerRPC_ListRems_FullMethodName, in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *listenerRPCClient) NewProfile(ctx context.Context, in *clientpb.Profile, opts ...grpc.CallOption) (*clientpb.Empty, error) {
-	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_NewProfile_FullMethodName, in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *listenerRPCClient) FindArtifact(ctx context.Context, in *clientpb.Artifact, opts ...grpc.CallOption) (*clientpb.Artifact, error) {
-	out := new(clientpb.Artifact)
-	err := c.cc.Invoke(ctx, ListenerRPC_FindArtifact_FullMethodName, in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *listenerRPCClient) Build(ctx context.Context, in *clientpb.Generate, opts ...grpc.CallOption) (*clientpb.Builder, error) {
-	out := new(clientpb.Builder)
-	err := c.cc.Invoke(ctx, ListenerRPC_Build_FullMethodName, in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *listenerRPCClient) DockerStatus(ctx context.Context, in *clientpb.Empty, opts ...grpc.CallOption) (*clientpb.Empty, error) {
-	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_DockerStatus_FullMethodName, in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *listenerRPCClient) TriggerWorkflowDispatch(ctx context.Context, in *clientpb.GithubWorkflowRequest, opts ...grpc.CallOption) (*clientpb.Builder, error) {
-	out := new(clientpb.Builder)
-	err := c.cc.Invoke(ctx, ListenerRPC_TriggerWorkflowDispatch_FullMethodName, in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *listenerRPCClient) WorkflowStatus(ctx context.Context, in *clientpb.GithubWorkflowRequest, opts ...grpc.CallOption) (*clientpb.Empty, error) {
-	out := new(clientpb.Empty)
-	err := c.cc.Invoke(ctx, ListenerRPC_WorkflowStatus_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_ListRems_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -442,8 +314,9 @@ func (c *listenerRPCClient) WorkflowStatus(ctx context.Context, in *clientpb.Git
 }
 
 func (c *listenerRPCClient) GetArtifact(ctx context.Context, in *clientpb.Artifact, opts ...grpc.CallOption) (*clientpb.Artifact, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(clientpb.Artifact)
-	err := c.cc.Invoke(ctx, ListenerRPC_GetArtifact_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ListenerRPC_GetArtifact_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -452,10 +325,10 @@ func (c *listenerRPCClient) GetArtifact(ctx context.Context, in *clientpb.Artifa
 
 // ListenerRPCServer is the server API for ListenerRPC service.
 // All implementations must embed UnimplementedListenerRPCServer
-// for forward compatibility
+// for forward compatibility.
 type ListenerRPCServer interface {
-	SpiteStream(ListenerRPC_SpiteStreamServer) error
-	JobStream(ListenerRPC_JobStreamServer) error
+	SpiteStream(grpc.BidiStreamingServer[clientpb.SpiteResponse, clientpb.SpiteRequest]) error
+	JobStream(grpc.BidiStreamingServer[clientpb.JobStatus, clientpb.JobCtrl]) error
 	// implant
 	Register(context.Context, *clientpb.RegisterSession) (*clientpb.Empty, error)
 	SysInfo(context.Context, *implantpb.SysInfo) (*clientpb.Empty, error)
@@ -469,41 +342,32 @@ type ListenerRPCServer interface {
 	DeletePipeline(context.Context, *clientpb.CtrlPipeline) (*clientpb.Empty, error)
 	ListPipelines(context.Context, *clientpb.Listener) (*clientpb.Pipelines, error)
 	// website
-	RegisterWebsite(context.Context, *clientpb.Pipeline) (*clientpb.Empty, error)
+	RegisterWebsite(context.Context, *clientpb.Pipeline) (*clientpb.WebsiteResponse, error)
 	StartWebsite(context.Context, *clientpb.CtrlPipeline) (*clientpb.Empty, error)
 	StopWebsite(context.Context, *clientpb.CtrlPipeline) (*clientpb.Empty, error)
-	DeleteWebsite(context.Context, *clientpb.CtrlPipeline) (*clientpb.Empty, error)
+	UploadWebsite(context.Context, *clientpb.Website) (*clientpb.Empty, error)
 	ListWebsites(context.Context, *clientpb.Listener) (*clientpb.Pipelines, error)
-	ListWebContent(context.Context, *clientpb.Website) (*clientpb.WebContents, error)
-	WebsiteAddContent(context.Context, *clientpb.Website) (*clientpb.Empty, error)
-	WebsiteUpdateContent(context.Context, *clientpb.WebContent) (*clientpb.Empty, error)
-	WebsiteRemoveContent(context.Context, *clientpb.WebContent) (*clientpb.Empty, error)
 	// rem
 	RegisterRem(context.Context, *clientpb.Pipeline) (*clientpb.Empty, error)
 	StartRem(context.Context, *clientpb.CtrlPipeline) (*clientpb.Empty, error)
 	StopRem(context.Context, *clientpb.CtrlPipeline) (*clientpb.Empty, error)
 	DeleteRem(context.Context, *clientpb.CtrlPipeline) (*clientpb.Empty, error)
 	ListRems(context.Context, *clientpb.Listener) (*clientpb.Pipelines, error)
-	// generator
-	NewProfile(context.Context, *clientpb.Profile) (*clientpb.Empty, error)
-	FindArtifact(context.Context, *clientpb.Artifact) (*clientpb.Artifact, error)
-	Build(context.Context, *clientpb.Generate) (*clientpb.Builder, error)
-	DockerStatus(context.Context, *clientpb.Empty) (*clientpb.Empty, error)
-	// action
-	TriggerWorkflowDispatch(context.Context, *clientpb.GithubWorkflowRequest) (*clientpb.Builder, error)
-	WorkflowStatus(context.Context, *clientpb.GithubWorkflowRequest) (*clientpb.Empty, error)
 	GetArtifact(context.Context, *clientpb.Artifact) (*clientpb.Artifact, error)
 	mustEmbedUnimplementedListenerRPCServer()
 }
 
-// UnimplementedListenerRPCServer must be embedded to have forward compatible implementations.
-type UnimplementedListenerRPCServer struct {
-}
+// UnimplementedListenerRPCServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedListenerRPCServer struct{}
 
-func (UnimplementedListenerRPCServer) SpiteStream(ListenerRPC_SpiteStreamServer) error {
+func (UnimplementedListenerRPCServer) SpiteStream(grpc.BidiStreamingServer[clientpb.SpiteResponse, clientpb.SpiteRequest]) error {
 	return status.Errorf(codes.Unimplemented, "method SpiteStream not implemented")
 }
-func (UnimplementedListenerRPCServer) JobStream(ListenerRPC_JobStreamServer) error {
+func (UnimplementedListenerRPCServer) JobStream(grpc.BidiStreamingServer[clientpb.JobStatus, clientpb.JobCtrl]) error {
 	return status.Errorf(codes.Unimplemented, "method JobStream not implemented")
 }
 func (UnimplementedListenerRPCServer) Register(context.Context, *clientpb.RegisterSession) (*clientpb.Empty, error) {
@@ -536,7 +400,7 @@ func (UnimplementedListenerRPCServer) DeletePipeline(context.Context, *clientpb.
 func (UnimplementedListenerRPCServer) ListPipelines(context.Context, *clientpb.Listener) (*clientpb.Pipelines, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListPipelines not implemented")
 }
-func (UnimplementedListenerRPCServer) RegisterWebsite(context.Context, *clientpb.Pipeline) (*clientpb.Empty, error) {
+func (UnimplementedListenerRPCServer) RegisterWebsite(context.Context, *clientpb.Pipeline) (*clientpb.WebsiteResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method RegisterWebsite not implemented")
 }
 func (UnimplementedListenerRPCServer) StartWebsite(context.Context, *clientpb.CtrlPipeline) (*clientpb.Empty, error) {
@@ -545,23 +409,11 @@ func (UnimplementedListenerRPCServer) StartWebsite(context.Context, *clientpb.Ct
 func (UnimplementedListenerRPCServer) StopWebsite(context.Context, *clientpb.CtrlPipeline) (*clientpb.Empty, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method StopWebsite not implemented")
 }
-func (UnimplementedListenerRPCServer) DeleteWebsite(context.Context, *clientpb.CtrlPipeline) (*clientpb.Empty, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method DeleteWebsite not implemented")
+func (UnimplementedListenerRPCServer) UploadWebsite(context.Context, *clientpb.Website) (*clientpb.Empty, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method UploadWebsite not implemented")
 }
 func (UnimplementedListenerRPCServer) ListWebsites(context.Context, *clientpb.Listener) (*clientpb.Pipelines, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListWebsites not implemented")
-}
-func (UnimplementedListenerRPCServer) ListWebContent(context.Context, *clientpb.Website) (*clientpb.WebContents, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method ListWebContent not implemented")
-}
-func (UnimplementedListenerRPCServer) WebsiteAddContent(context.Context, *clientpb.Website) (*clientpb.Empty, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method WebsiteAddContent not implemented")
-}
-func (UnimplementedListenerRPCServer) WebsiteUpdateContent(context.Context, *clientpb.WebContent) (*clientpb.Empty, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method WebsiteUpdateContent not implemented")
-}
-func (UnimplementedListenerRPCServer) WebsiteRemoveContent(context.Context, *clientpb.WebContent) (*clientpb.Empty, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method WebsiteRemoveContent not implemented")
 }
 func (UnimplementedListenerRPCServer) RegisterRem(context.Context, *clientpb.Pipeline) (*clientpb.Empty, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method RegisterRem not implemented")
@@ -578,28 +430,11 @@ func (UnimplementedListenerRPCServer) DeleteRem(context.Context, *clientpb.CtrlP
 func (UnimplementedListenerRPCServer) ListRems(context.Context, *clientpb.Listener) (*clientpb.Pipelines, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListRems not implemented")
 }
-func (UnimplementedListenerRPCServer) NewProfile(context.Context, *clientpb.Profile) (*clientpb.Empty, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method NewProfile not implemented")
-}
-func (UnimplementedListenerRPCServer) FindArtifact(context.Context, *clientpb.Artifact) (*clientpb.Artifact, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method FindArtifact not implemented")
-}
-func (UnimplementedListenerRPCServer) Build(context.Context, *clientpb.Generate) (*clientpb.Builder, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method Build not implemented")
-}
-func (UnimplementedListenerRPCServer) DockerStatus(context.Context, *clientpb.Empty) (*clientpb.Empty, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method DockerStatus not implemented")
-}
-func (UnimplementedListenerRPCServer) TriggerWorkflowDispatch(context.Context, *clientpb.GithubWorkflowRequest) (*clientpb.Builder, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method TriggerWorkflowDispatch not implemented")
-}
-func (UnimplementedListenerRPCServer) WorkflowStatus(context.Context, *clientpb.GithubWorkflowRequest) (*clientpb.Empty, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method WorkflowStatus not implemented")
-}
 func (UnimplementedListenerRPCServer) GetArtifact(context.Context, *clientpb.Artifact) (*clientpb.Artifact, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetArtifact not implemented")
 }
 func (UnimplementedListenerRPCServer) mustEmbedUnimplementedListenerRPCServer() {}
+func (UnimplementedListenerRPCServer) testEmbeddedByValue()                     {}
 
 // UnsafeListenerRPCServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ListenerRPCServer will
@@ -609,60 +444,29 @@ type UnsafeListenerRPCServer interface {
 }
 
 func RegisterListenerRPCServer(s grpc.ServiceRegistrar, srv ListenerRPCServer) {
+	// If the following call pancis, it indicates UnimplementedListenerRPCServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&ListenerRPC_ServiceDesc, srv)
 }
 
 func _ListenerRPC_SpiteStream_Handler(srv interface{}, stream grpc.ServerStream) error {
-	return srv.(ListenerRPCServer).SpiteStream(&listenerRPCSpiteStreamServer{stream})
+	return srv.(ListenerRPCServer).SpiteStream(&grpc.GenericServerStream[clientpb.SpiteResponse, clientpb.SpiteRequest]{ServerStream: stream})
 }
 
-type ListenerRPC_SpiteStreamServer interface {
-	Send(*clientpb.SpiteRequest) error
-	Recv() (*clientpb.SpiteResponse, error)
-	grpc.ServerStream
-}
-
-type listenerRPCSpiteStreamServer struct {
-	grpc.ServerStream
-}
-
-func (x *listenerRPCSpiteStreamServer) Send(m *clientpb.SpiteRequest) error {
-	return x.ServerStream.SendMsg(m)
-}
-
-func (x *listenerRPCSpiteStreamServer) Recv() (*clientpb.SpiteResponse, error) {
-	m := new(clientpb.SpiteResponse)
-	if err := x.ServerStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ListenerRPC_SpiteStreamServer = grpc.BidiStreamingServer[clientpb.SpiteResponse, clientpb.SpiteRequest]
 
 func _ListenerRPC_JobStream_Handler(srv interface{}, stream grpc.ServerStream) error {
-	return srv.(ListenerRPCServer).JobStream(&listenerRPCJobStreamServer{stream})
+	return srv.(ListenerRPCServer).JobStream(&grpc.GenericServerStream[clientpb.JobStatus, clientpb.JobCtrl]{ServerStream: stream})
 }
 
-type ListenerRPC_JobStreamServer interface {
-	Send(*clientpb.JobCtrl) error
-	Recv() (*clientpb.JobStatus, error)
-	grpc.ServerStream
-}
-
-type listenerRPCJobStreamServer struct {
-	grpc.ServerStream
-}
-
-func (x *listenerRPCJobStreamServer) Send(m *clientpb.JobCtrl) error {
-	return x.ServerStream.SendMsg(m)
-}
-
-func (x *listenerRPCJobStreamServer) Recv() (*clientpb.JobStatus, error) {
-	m := new(clientpb.JobStatus)
-	if err := x.ServerStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ListenerRPC_JobStreamServer = grpc.BidiStreamingServer[clientpb.JobStatus, clientpb.JobCtrl]
 
 func _ListenerRPC_Register_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(clientpb.RegisterSession)
@@ -898,20 +702,20 @@ func _ListenerRPC_StopWebsite_Handler(srv interface{}, ctx context.Context, dec 
 	return interceptor(ctx, in, info, handler)
 }
 
-func _ListenerRPC_DeleteWebsite_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(clientpb.CtrlPipeline)
+func _ListenerRPC_UploadWebsite_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(clientpb.Website)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(ListenerRPCServer).DeleteWebsite(ctx, in)
+		return srv.(ListenerRPCServer).UploadWebsite(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: ListenerRPC_DeleteWebsite_FullMethodName,
+		FullMethod: ListenerRPC_UploadWebsite_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(ListenerRPCServer).DeleteWebsite(ctx, req.(*clientpb.CtrlPipeline))
+		return srv.(ListenerRPCServer).UploadWebsite(ctx, req.(*clientpb.Website))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -930,78 +734,6 @@ func _ListenerRPC_ListWebsites_Handler(srv interface{}, ctx context.Context, dec
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ListenerRPCServer).ListWebsites(ctx, req.(*clientpb.Listener))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _ListenerRPC_ListWebContent_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(clientpb.Website)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(ListenerRPCServer).ListWebContent(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: ListenerRPC_ListWebContent_FullMethodName,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(ListenerRPCServer).ListWebContent(ctx, req.(*clientpb.Website))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _ListenerRPC_WebsiteAddContent_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(clientpb.Website)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(ListenerRPCServer).WebsiteAddContent(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: ListenerRPC_WebsiteAddContent_FullMethodName,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(ListenerRPCServer).WebsiteAddContent(ctx, req.(*clientpb.Website))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _ListenerRPC_WebsiteUpdateContent_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(clientpb.WebContent)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(ListenerRPCServer).WebsiteUpdateContent(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: ListenerRPC_WebsiteUpdateContent_FullMethodName,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(ListenerRPCServer).WebsiteUpdateContent(ctx, req.(*clientpb.WebContent))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _ListenerRPC_WebsiteRemoveContent_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(clientpb.WebContent)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(ListenerRPCServer).WebsiteRemoveContent(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: ListenerRPC_WebsiteRemoveContent_FullMethodName,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(ListenerRPCServer).WebsiteRemoveContent(ctx, req.(*clientpb.WebContent))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -1096,114 +828,6 @@ func _ListenerRPC_ListRems_Handler(srv interface{}, ctx context.Context, dec fun
 	return interceptor(ctx, in, info, handler)
 }
 
-func _ListenerRPC_NewProfile_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(clientpb.Profile)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(ListenerRPCServer).NewProfile(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: ListenerRPC_NewProfile_FullMethodName,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(ListenerRPCServer).NewProfile(ctx, req.(*clientpb.Profile))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _ListenerRPC_FindArtifact_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(clientpb.Artifact)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(ListenerRPCServer).FindArtifact(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: ListenerRPC_FindArtifact_FullMethodName,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(ListenerRPCServer).FindArtifact(ctx, req.(*clientpb.Artifact))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _ListenerRPC_Build_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(clientpb.Generate)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(ListenerRPCServer).Build(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: ListenerRPC_Build_FullMethodName,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(ListenerRPCServer).Build(ctx, req.(*clientpb.Generate))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _ListenerRPC_DockerStatus_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(clientpb.Empty)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(ListenerRPCServer).DockerStatus(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: ListenerRPC_DockerStatus_FullMethodName,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(ListenerRPCServer).DockerStatus(ctx, req.(*clientpb.Empty))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _ListenerRPC_TriggerWorkflowDispatch_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(clientpb.GithubWorkflowRequest)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(ListenerRPCServer).TriggerWorkflowDispatch(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: ListenerRPC_TriggerWorkflowDispatch_FullMethodName,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(ListenerRPCServer).TriggerWorkflowDispatch(ctx, req.(*clientpb.GithubWorkflowRequest))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _ListenerRPC_WorkflowStatus_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(clientpb.GithubWorkflowRequest)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(ListenerRPCServer).WorkflowStatus(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: ListenerRPC_WorkflowStatus_FullMethodName,
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(ListenerRPCServer).WorkflowStatus(ctx, req.(*clientpb.GithubWorkflowRequest))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
 func _ListenerRPC_GetArtifact_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(clientpb.Artifact)
 	if err := dec(in); err != nil {
@@ -1282,28 +906,12 @@ var ListenerRPC_ServiceDesc = grpc.ServiceDesc{
 			Handler:    _ListenerRPC_StopWebsite_Handler,
 		},
 		{
-			MethodName: "DeleteWebsite",
-			Handler:    _ListenerRPC_DeleteWebsite_Handler,
+			MethodName: "UploadWebsite",
+			Handler:    _ListenerRPC_UploadWebsite_Handler,
 		},
 		{
 			MethodName: "ListWebsites",
 			Handler:    _ListenerRPC_ListWebsites_Handler,
-		},
-		{
-			MethodName: "ListWebContent",
-			Handler:    _ListenerRPC_ListWebContent_Handler,
-		},
-		{
-			MethodName: "WebsiteAddContent",
-			Handler:    _ListenerRPC_WebsiteAddContent_Handler,
-		},
-		{
-			MethodName: "WebsiteUpdateContent",
-			Handler:    _ListenerRPC_WebsiteUpdateContent_Handler,
-		},
-		{
-			MethodName: "WebsiteRemoveContent",
-			Handler:    _ListenerRPC_WebsiteRemoveContent_Handler,
 		},
 		{
 			MethodName: "RegisterRem",
@@ -1324,30 +932,6 @@ var ListenerRPC_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "ListRems",
 			Handler:    _ListenerRPC_ListRems_Handler,
-		},
-		{
-			MethodName: "NewProfile",
-			Handler:    _ListenerRPC_NewProfile_Handler,
-		},
-		{
-			MethodName: "FindArtifact",
-			Handler:    _ListenerRPC_FindArtifact_Handler,
-		},
-		{
-			MethodName: "Build",
-			Handler:    _ListenerRPC_Build_Handler,
-		},
-		{
-			MethodName: "DockerStatus",
-			Handler:    _ListenerRPC_DockerStatus_Handler,
-		},
-		{
-			MethodName: "TriggerWorkflowDispatch",
-			Handler:    _ListenerRPC_TriggerWorkflowDispatch_Handler,
-		},
-		{
-			MethodName: "WorkflowStatus",
-			Handler:    _ListenerRPC_WorkflowStatus_Handler,
 		},
 		{
 			MethodName: "GetArtifact",
